refactor: Base64Convertor util 패키지로 이동
