+++ src/main/java/nextstep/security/filter/BasicAuthenticationFilter.java
@@ -8,7 +8,7 @@ import jakarta.servlet.ServletResponse;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.security.AuthenticationException;
-import nextstep.security.Base64Convertor;
+import nextstep.security.util.Base64Convertor;
 import nextstep.security.UserDetails;
 import nextstep.security.UserDetailsService;
